Add TranslatableLanguage to BCP47LanguageIdentifier
